change configurations

+++ Luqma.Data/Entities/Cart.cs
@@ -13,8 +13,6 @@ namespace Luqma.Data.Entities
         public virtual Customer Customer { get; set; }
         public int ItemId { get; set; }
         public virtual MenuItem MenuItem { get; set; }
-        public int Quantity { get; set; }
-
-
+        public double Quantity { get; set; }
     }
 }
