+++ Luqma.Data/Response/Carts/CartResponse.cs
@@ -12,9 +12,6 @@ namespace Luqma.Data.Response.Carts
         public string? ImageUrl { get; set; }
         public string? Description { get; set; }
         public double Price { get; set; }
-        public int Quantity { get; set; }
-
-
-
+        public double Quantity { get; set; }
     }
 }
